* feature/full_usd: (25 commits)
  feat: Remove unneeded gearbox-core, -physics, -viz crates
  feat: Implement session-based command velocity control
  feat: Improve visual wheel spin for Ackermann steering
  feat: Implement Ackermann steering differential
  feat: Add per-axle steering multiplier
  feat: Add Oxbo pea harvester capabilities
  feat: Integrate USD-based terrain for enhanced worlds
  feat: robot-usd
  feat: Add USD pose feedback and marker API
  feat: add local terrain mesh with physics and raycast vehicle
  feat: Implement fully physical wheel simulation
  feat: Rework simulator to use new USD pipeline
  feat(editor): USDs section in workspace tree
  feat(editor): inspector shows USD world pose + geo
  feat(editor): selection ring extends to USD entities
  feat(editor): teleport USD bodies to visuals while paused
  feat(editor): Load USD ribbon + USD selection + gizmo on USD roots
  feat(editor): USD support via --usd, mirrors play/pause
  feat(world): bevy_frost ribbon + transport panels
  feat(world): grid layout + AnimPlugin wiring
  ...

+++ crates/gearbox-core/src/presets/drone.rs
@@ -14,8 +14,8 @@
 use datapod::{Point, Quaternion, Size};
 
 use crate::vehicle::{
-    ChassisSpec, DriveMode, MeshSource, PartKind, PartSpec, PowerKind, PowerSource,
-    VehicleBuilder, VehicleSpec,
+    ChassisSpec, DriveMode, MeshSource, PartKind, PartSpec, PowerKind, PowerSource, VehicleBuilder,
+    VehicleSpec,
 };
 
 pub fn drone() -> VehicleSpec {
@@ -55,10 +55,10 @@ pub fn drone() -> VehicleSpec {
     // Four arms extend diagonally from the body's corners; a small
     // flat disc at the tip reads as a spinning rotor. Visual-only —
     // the flight model applies forces at the CoM, not per-rotor.
-    let arm_len = 0.18;   // from body corner out to rotor hub
-    let arm_h   = 0.03;
-    let arm_w   = 0.04;
-    let rotor_r = 0.10;   // rotor disc radius (size.x == size.z)
+    let arm_len = 0.18; // from body corner out to rotor hub
+    let arm_h = 0.03;
+    let arm_w = 0.04;
+    let rotor_r = 0.10; // rotor disc radius (size.x == size.z)
 
     let body_half_x = (chassis_x as f32) * 0.5;
     let body_half_z = (chassis_z as f32) * 0.5;
@@ -68,7 +68,7 @@ pub fn drone() -> VehicleSpec {
     // Arms stay a dark neutral so the orange body pops; rotor discs
     // are kept bright but intentionally a different hue from the
     // chassis so the props read clearly while spinning.
-    let arm_color   = [0.08, 0.08, 0.10];
+    let arm_color = [0.08, 0.08, 0.10];
     let rotor_color = [0.90, 0.90, 0.92];
 
     let make_arm = |name: &str, dir_x: f32, dir_z: f32| PartSpec {
@@ -78,7 +78,7 @@ pub fn drone() -> VehicleSpec {
             0.0,
             (body_half_z * dir_z.signum() + diag_off * dir_z * 0.5) as f64,
         ),
-        size: Size::new(arm_w as f64, arm_h as f64, arm_len as f64),
+        size: Size::new(arm_w, arm_h, arm_len as f64),
         color: arm_color,
         // `Hitch` → visual-only: no rapier collider is built for this
         // part, only its Bevy mesh. Necessary for the drone because:
@@ -99,7 +99,7 @@ pub fn drone() -> VehicleSpec {
             rotor_y as f64,
             (dir_z * (body_half_z + diag_off)) as f64,
         ),
-        size: Size::new((rotor_r * 2.0) as f64, 0.01, (rotor_r * 2.0) as f64),
+        size: Size::new(rotor_r * 2.0, 0.01, rotor_r * 2.0),
         color: rotor_color,
         kind: PartKind::Hitch, // visual-only (see note above)
         mesh: MeshSource::Box,
@@ -108,14 +108,14 @@ pub fn drone() -> VehicleSpec {
     VehicleBuilder::new("drone", chassis)
         .max_speed(15.0)
         // Four diagonal arms.
-        .part(make_arm("arm_fr",  1.0,  1.0))
-        .part(make_arm("arm_fl", -1.0,  1.0))
-        .part(make_arm("arm_rr",  1.0, -1.0))
+        .part(make_arm("arm_fr", 1.0, 1.0))
+        .part(make_arm("arm_fl", -1.0, 1.0))
+        .part(make_arm("arm_rr", 1.0, -1.0))
         .part(make_arm("arm_rl", -1.0, -1.0))
         // Four rotor discs at the arm tips.
-        .part(make_rotor("rotor_fr",  1.0,  1.0))
-        .part(make_rotor("rotor_fl", -1.0,  1.0))
-        .part(make_rotor("rotor_rr",  1.0, -1.0))
+        .part(make_rotor("rotor_fr", 1.0, 1.0))
+        .part(make_rotor("rotor_fl", -1.0, 1.0))
+        .part(make_rotor("rotor_rr", 1.0, -1.0))
         .part(make_rotor("rotor_rl", -1.0, -1.0))
         .drive_mode(DriveMode::Drone)
         .power_source(
